// sdk/search/search-documents/generated/models/azure/search/documents/indexes/models.d.ts
export interface SynonymMap {
    /** The name of the synonym map. */
    name: string;
    /** The format of the synonym map. Only the 'solr' format is currently supported. */
    format: "solr";
    /** A series of synonym rules in the specified synonym map format. The rules must be separated by newlines. */
    synonyms: string[];
    /** A description of an encryption key that you create in Azure Key Vault. This key is used to provide an additional level of encryption-at-rest for your data when you want full assurance that no one, not even Microsoft, can decrypt your data. Once you have encrypted your data, it will always remain encrypted. The search service will ignore attempts to set this property to null. You can change this property as needed if you want to rotate your encryption key; Your data will be unaffected. Encryption with customer-managed keys is not available for free search services, and is only available for paid services created on or after January 1, 2019. */
    encryptionKey?: SearchResourceEncryptionKey;
    /** The ETag of the synonym map. */
    eTag?: string;
}
// ...
export interface ListSynonymMapsResult {
    /** The synonym maps in the Search service. */
    readonly synonymMaps: SynonymMap[];
}
export declare function listSynonymMapsResultDeserializer(item: any): ListSynonymMapsResult;
export declare function synonymMapArraySerializer(result: Array<SynonymMap>): any[];
export declare function synonymMapArrayDeserializer(result: Array<SynonymMap>): any[];
/** Represents a search index definition, which describes the fields and search behavior of an index. */
export interface SearchIndex {
    /** The name of the index. */
    name: string;
    /** The description of the index. */
    description?: string;
    /** The fields of the index. */
    fields: SearchField[];
    /** The scoring profiles for the index. */
    scoringProfiles?: ScoringProfile[];
    /** The name of the scoring profile to use if none is specified in the query. If this property is not set and no scoring profile is specified in the query, then default scoring (tf-idf) will be used. */
    defaultScoringProfile?: string;
    /** Options to control Cross-Origin Resource Sharing (CORS) for the index. */
    corsOptions?: CorsOptions;
    /** The suggesters for the index. */
    suggesters?: SearchSuggester[];
    /** The analyzers for the index. */
    analyzers?: LexicalAnalyzerUnion[];
    /** The tokenizers for the index. */
    tokenizers?: LexicalTokenizerUnion[];
    /** The token filters for the index. */
    tokenFilters?: TokenFilterUnion[];
    /** The character filters for the index. */
    charFilters?: CharFilterUnion[];
    /** The normalizers for the index. */
    normalizers?: LexicalNormalizerUnion[];
    /** A description of an encryption key that you create in Azure Key Vault. This key is used to provide an additional level of encryption-at-rest for your data when you want full assurance that no one, not even Microsoft, can decrypt your data. Once you have encrypted your data, it will always remain encrypted. The search service will ignore attempts to set this property to null. You can change this property as needed if you want to rotate your encryption key; Your data will be unaffected. Encryption with customer-managed keys is not available for free search services, and is only available for paid services created on or after January 1, 2019. */
    encryptionKey?: SearchResourceEncryptionKey;
    /** The type of similarity algorithm to be used when scoring and ranking the documents matching a search query. The similarity algorithm can only be defined at index creation time and cannot be modified on existing indexes. If null, the ClassicSimilarity algorithm is used. */
    similarity?: SimilarityAlgorithmUnion;
    /** Defines parameters for a search index that influence semantic capabilities. */
    semanticSearch?: SemanticSearch;
    /** Contains configuration options related to vector search. */
    vectorSearch?: VectorSearch;
    /** A value indicating whether permission filtering is enabled for the index. */
    permissionFilterOption?: SearchIndexPermissionFilterOption;
    /** A value indicating whether Purview is enabled for the index. */
    purviewEnabled?: boolean;
    /** Configures a SharePoint connector app registration for the index, enabling document-level permissions from SharePoint. If provided, the applicationId and federatedCredentialId properties are required. */
    sharePointConnectorAppRegistration?: SharePointConnectorAppRegistration;
    /** The ETag of the index. */
    eTag?: string;
}
// ...
export interface ScoringProfile {
    /** The name of the scoring profile. */
    name: string;
    /** Parameters that boost scoring based on text matches in certain index fields. */
    textWeights?: TextWeights;
    /** The collection of functions that influence the scoring of documents. */
    functions?: ScoringFunctionUnion[];
    /** A value indicating how the results of individual scoring functions should be combined. Defaults to "Sum". Ignored if there are no scoring functions. */
    functionAggregation?: ScoringFunctionAggregation;
}
// ...
export interface SearchSuggester {
    /** The name of the suggester. */
    name: string;
    /** A value indicating the capabilities of the suggester. */
    searchMode: "analyzingInfixMatching";
    /** The list of field names to which the suggester applies. Each field must be searchable. */
    sourceFields: string[];
}
// ...
export type CharFilterUnion = MappingCharFilter | PatternReplaceCharFilter | CharFilter;
// ...
export interface SearchIndexResponse {
    /** The name of the index. */
    name: string;
    /** The description of the index. */
    description?: string;
    /** The fields of the index. */
    fields?: SearchField[];
    /** The scoring profiles for the index. */
    scoringProfiles?: ScoringProfile[];
    /** The name of the scoring profile to use if none is specified in the query. If this property is not set and no scoring profile is specified in the query, then default scoring (tf-idf) will be used. */
    defaultScoringProfile?: string;
    /** Options to control Cross-Origin Resource Sharing (CORS) for the index. */
    corsOptions?: CorsOptions;
    /** The suggesters for the index. */
    suggesters?: SearchSuggester[];
    /** The analyzers for the index. */
    analyzers?: LexicalAnalyzerUnion[];
    /** The tokenizers for the index. */
    tokenizers?: LexicalTokenizerUnion[];
    /** The token filters for the index. */
    tokenFilters?: TokenFilterUnion[];
    /** The character filters for the index. */
    charFilters?: CharFilterUnion[];
    /** The normalizers for the index. */
    normalizers?: LexicalNormalizerUnion[];
    /** A description of an encryption key that you create in Azure Key Vault. This key is used to provide an additional level of encryption-at-rest for your data when you want full assurance that no one, not even Microsoft, can decrypt your data. Once you have encrypted your data, it will always remain encrypted. The search service will ignore attempts to set this property to null. You can change this property as needed if you want to rotate your encryption key; Your data will be unaffected. Encryption with customer-managed keys is not available for free search services, and is only available for paid services created on or after January 1, 2019. */
    encryptionKey?: SearchResourceEncryptionKey;
    /** The type of similarity algorithm to be used when scoring and ranking the documents matching a search query. The similarity algorithm can only be defined at index creation time and cannot be modified on existing indexes. If null, the ClassicSimilarity algorithm is used. */
    similarity?: SimilarityAlgorithmUnion;
    /** Defines parameters for a search index that influence semantic capabilities. */
    semanticSearch?: SemanticSearch;
    /** Contains configuration options related to vector search. */
    vectorSearch?: VectorSearch;
    /** A value indicating whether permission filtering is enabled for the index. */
    permissionFilterOption?: SearchIndexPermissionFilterOption;
    /** A value indicating whether Purview is enabled for the index. */
    purviewEnabled?: boolean;
    /** The ETag of the index. */
    eTag?: string;
}
export declare function searchIndexResponseDeserializer(item: any): SearchIndexResponse;
/** Statistics for a given index. Statistics are collected periodically and are not guaranteed to always be up-to-date. */
export interface GetIndexStatisticsResult {
    /** The number of documents in the index. */
    readonly documentCount: number;
    /** The amount of storage in bytes consumed by the index. */
    readonly storageSize: number;
    /** The amount of memory in bytes consumed by vectors in the index. */
    readonly vectorIndexSize: number;
}
export declare function getIndexStatisticsResultSerializer(_item: GetIndexStatisticsResult): any;
export declare function getIndexStatisticsResultDeserializer(item: any): GetIndexStatisticsResult;
/** Specifies some text and analysis components used to break that text into tokens. */
export interface AnalyzeTextOptions {
    /** The text to break into tokens. */
    text: string;
    /** The name of the analyzer to use to break the given text. If this parameter is not specified, you must specify a tokenizer instead. The tokenizer and analyzer parameters are mutually exclusive. */
    analyzerName?: LexicalAnalyzerName;
    /** The name of the tokenizer to use to break the given text. If this parameter is not specified, you must specify an analyzer instead. The tokenizer and analyzer parameters are mutually exclusive. */
    tokenizerName?: LexicalTokenizerName;
    /** The name of the normalizer to use to normalize the given text. */
    normalizerName?: LexicalNormalizerName;
    /** An optional list of token filters to use when breaking the given text. This parameter can only be set when using the tokenizer parameter. */
    tokenFilters?: TokenFilterName[];
    /** An optional list of character filters to use when breaking the given text. This parameter can only be set when using the tokenizer parameter. */
    charFilters?: CharFilterName[];
}
export declare function analyzeTextOptionsSerializer(item: AnalyzeTextOptions): any;
/** The result of testing an analyzer on text. */
export interface AnalyzeResult {
    /** The list of tokens returned by the analyzer specified in the request. */
    tokens: AnalyzedTokenInfo[];
}
// ...
export interface SearchAlias {
    /** The name of the alias. */
    name: string;
    /** The name of the index this alias maps to. Only one index name may be specified. */
    indexes: string[];
    /** The ETag of the alias. */
    etag?: string;
}
// ...
export interface KnowledgeBase {
    /** The name of the knowledge base. */
    name: string;
    /** Knowledge sources referenced by this knowledge base. */
    knowledgeSources: KnowledgeSourceReference[];
    /** Contains configuration options on how to connect to AI models. */
    models?: KnowledgeBaseModelUnion[];
    /** The retrieval reasoning effort configuration. */
    retrievalReasoningEffort?: KnowledgeRetrievalReasoningEffortUnion;
    /** The output mode for the knowledge base. */
    outputMode?: KnowledgeRetrievalOutputMode;
    /** The ETag of the knowledge base. */
    etag?: string;
    /** A description of an encryption key that you create in Azure Key Vault. */
    encryptionKey?: SearchResourceEncryptionKey;
    /** The description of the knowledge base. */
    description?: string;
    /** Instructions considered by the knowledge base when developing query plan. */
    retrievalInstructions?: string;
    /** Instructions considered by the knowledge base when generating answers. */
    answerInstructions?: string;
    /** Options to control Cross-Origin Resource Sharing (CORS) for the knowledge base. */
    corsOptions?: CorsOptions;
}
// ...
export type KnowledgeSourceUnion = SearchIndexKnowledgeSource | AzureBlobKnowledgeSource | IndexedSharePointKnowledgeSource | IndexedOneLakeKnowledgeSource | IndexedSqlKnowledgeSource | FileKnowledgeSource | WebKnowledgeSource | RemoteSharePointKnowledgeSource | WorkIQKnowledgeSource | McpServerKnowledgeSource | FabricDataAgentKnowledgeSource | FabricOntologyKnowledgeSource | KnowledgeSource;
// ...
export interface SearchServiceStatistics {
    /** Service level resource counters. */
    counters: ServiceCounters;
    /** Service level general limits. */
    limits: ServiceLimits;
    /** Service level indexer runtime consumption. */
    indexersRuntime: ServiceIndexersRuntime;
}
// ...
export interface IndexStatisticsSummary {
    /** The name of the index. */
    name: string;
    /** The number of documents in the index. */
    readonly documentCount: number;
    /** The amount of storage in bytes consumed by the index. */
    readonly storageSize: number;
    /** The amount of memory in bytes consumed by vectors in the index. */
    readonly vectorIndexSize: number;
}